+++ packages/lib/src/types.ts
@@ -1,3 +1,5 @@
+import type { HTMLButtonAttributes } from 'svelte/elements';
+
 export interface LoginProps {
   buttonText?: string;
   loading?: boolean;
@@ -11,3 +13,16 @@ export interface PaywallProps extends HTMLElement {
   ['settings-url']: string;
   template: 'ARTICLE' | 'BOXES';
 }
+
+export type Variant = 'primary' | 'secondary' | 'tertiary';
+export type Size = 'sm' | 'md' | 'lg';
+
+export type ButtonProps = HTMLButtonAttributes & {
+  loading?: boolean;
+  variant?: Variant;
+  disabled?: boolean;
+  size?: Size;
+  part?: string;
+  onclick?: () => void;
+  href?: string;
+};